Create ticket selectors and move from reducer
